Sending ID to the booking component from App.vue
I've set up a default userID as 0 in the main application (App.vue) and am passing that as a prop to the booking component. I now hope to be able to alter that default variable in the main app from the login component, and then send the update to the booking component. Obviously something needs to be put in place to ensure that user 0 doesn't make any bookings, but that validation can wait until I have a successful login -> booking scenario